# backend/src/main/java/com/tripscanner/TripScanner/controller/restController/PlaceManagementRestController.java
package com.tripscanner.TripScanner.controller.restController;

import com.tripscanner.TripScanner.model.Destination;
import com.tripscanner.TripScanner.model.Place;
import com.tripscanner.TripScanner.model.rest.PlaceDTO;
import com.tripscanner.TripScanner.service.DestinationService;
import com.tripscanner.TripScanner.service.ItineraryService;
import com.tripscanner.TripScanner.service.PlaceService;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.responses.ApiResponses;
import org.hibernate.engine.jdbc.BlobProxy;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.core.io.InputStreamResource;
import org.springframework.core.io.Resource;
import org.springframework.data.domain.Page;
import org.springframework.data.domain.PageRequest;
import org.springframework.http.HttpHeaders;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;
import org.springframework.web.servlet.support.ServletUriComponentsBuilder;

import javax.servlet.http.HttpServletRequest;
import java.io.IOException;
import java.net.URI;
import java.net.URISyntaxException;
import java.sql.SQLException;
import java.util.Optional;

import static org.springframework.web.servlet.support.ServletUriComponentsBuilder.fromCurrentRequest;

@RestController
@RequestMapping("/api/management/places")
public class PlaceManagementRestController {

    @Autowired
    private ItineraryService itineraryService;

    @Autowired
    private DestinationService destinationService;

    @Autowired
    private PlaceService placeService;

    @Operation(summary = "Get a page with all places")
    @ApiResponses(value = {
            @ApiResponse(
                    responseCode = "200",
                    description = "Places page",
                    content = {@Content(
                            mediaType = "application/json",
                            schema = @Schema(implementation= Place.class)
                    )}
            ),
            @ApiResponse(
                    responseCode = "403",
                    description = "Not authorized",
                    content = @Content
            ),
            @ApiResponse(
                    responseCode = "404",
                    description = "Places not found",
                    content = @Content
            )
    })

    @GetMapping("")
    public ResponseEntity<Page<Place>> getPlaces(
            @Parameter(description="page number")
            @RequestParam(defaultValue = "0") int page){
        Page<Place> places = placeService.findAll(PageRequest.of(page, 10));
        if (!places.isEmpty()) {
            return ResponseEntity.ok(places);
        } else {
            return ResponseEntity.notFound().build();
        }
    }

    @Operation(summary = "Create a new place")
    @ApiResponses(value = {
            @ApiResponse(
                    responseCode = "201",
                    description = "New place was created",
                    content = {@Content(
                            mediaType = "application/json",
                            schema = @Schema(implementation= Place.class)
                    )}
            ),
            @ApiResponse(
                    responseCode = "403",
                    description = "Not authorized",
                    content = @Content
            )
    })

    @PostMapping("")
    public ResponseEntity<Place> createPlace(
            @Parameter(description="new page's information")
            @RequestBody PlaceDTO newPlace) {
        Place place = new Place(newPlace.getName(), newPlace.getDescription(), destinationService.findByName(newPlace.getDestination()).get());
        placeService.save(place);
        URI location = fromCurrentRequest().path("/{id}").buildAndExpand(place.getId()).toUri();
        return ResponseEntity.created(location).body(place);
    }

    @Operation(summary = "Edit existing place")
    @ApiResponses(value = {
            @ApiResponse(
                    responseCode = "200",
                    description = "Place was correctly edited",
                    content = {@Content(
                            mediaType = "application/json",
                            schema = @Schema(implementation= Place.class)
                    )}
            ),
            @ApiResponse(
                    responseCode = "403",
                    description = "Not authorized",
                    content = @Content
            ),
            @ApiResponse(
                    responseCode = "404",
                    description = "Place not found",
                    content = @Content
            )
    })

    @PutMapping("/{id}")
    public ResponseEntity<Place> editPlace(
            @Parameter(description="id of place to be edited")
            @PathVariable long id,
            @Parameter(description="place's new information")
            @RequestBody PlaceDTO newPlaceDTO) throws SQLException {
        Place newPlace = new Place(newPlaceDTO.getName(), newPlaceDTO.getDescription(), destinationService.findByName(newPlaceDTO.getDestination()).get());
        Optional<Place> place = placeService.findById(id);

            if (place.isPresent()) {
                if (newPlace.getName() != null) place.get().setName(newPlace.getName());
                if (newPlace.getDescription() != null) place.get().setDescription(newPlace.getDescription());
                if (newPlace.getDestination() != null) {
                    place.get().setDestination(newPlace.getDestination());
                }
                place.get().setId(id);
                placeService.save(place.get());
                return new ResponseEntity(HttpStatus.OK);
        } else {
            return ResponseEntity.notFound().build();
        }
    }

    @Operation(summary = "Delete place")
    @ApiResponses(value = {
            @ApiResponse(
                    responseCode = "204",
                    description = "Place was correctly deleted",
                    content = @Content
            ),
            @ApiResponse(
                    responseCode = "403",
                    description = "Not authorized",
                    content = @Content
            ),
            @ApiResponse(
                    responseCode = "404",
                    description = "Place not found",
                    content = @Content
            )
    })

    @DeleteMapping("/{id}")
    public ResponseEntity<Place> deletePlace(
            @Parameter(description="id of place to be deleted")
            @PathVariable long id) {
        Optional<Place> place = placeService.findById(id);
        if (place.isPresent()) {
            for (int i = 0; i < place.get().getItineraries().size(); i++) {
                place.get().getItineraries().get(i).getPlaces().remove(place.get());
                itineraryService.findById(place.get().getItineraries().get(i).getId()).get().getPlaces().remove(place);
                if (itineraryService.findById(place.get().getItineraries().get(i).getId()).get().getPlaces().isEmpty()) {
                    itineraryService.delete(place.get().getItineraries().get(i).getId());
                }
            }
            placeService.delete(id);
            return new ResponseEntity<>(null, HttpStatus.NO_CONTENT);
        } else {
            return ResponseEntity.notFound().build();
        }
    }

    @Operation(summary = "Edit existing place's image")
    @ApiResponses(value = {
            @ApiResponse(
                    responseCode = "200",
                    description = "Place's image was edited correctly",
                    content = {@Content(
                            mediaType = "image/jpeg",
                            schema = @Schema(implementation= Place.class)
                    )}
            ),
            @ApiResponse(
                    responseCode = "403",
                    description = "Not authorized",
                    content = @Content
            ),
            @ApiResponse(
                    responseCode = "404",
                    description = "Place not found",
                    content = @Content
            )
    })

    @PutMapping("/{id}/image")
    public ResponseEntity<Resource> editImage(
            @Parameter(description="id of place to edit its image")
            @PathVariable long id,
            @Parameter(description="image to be uploaded")
            @RequestParam MultipartFile imageFile, HttpServletRequest request) throws IOException, URISyntaxException, SQLException {
        Optional<Place> place = placeService.findById(id);

        if (place.isPresent()) {
            Place newPlace = place.get();
            newPlace.setImageFile(BlobProxy.generateProxy(imageFile.getInputStream(), imageFile.getSize()));
            newPlace.setImage(true);
            placeService.save(newPlace);
            String baseUrl = ServletUriComponentsBuilder.fromRequestUri(request).replacePath(null).build().toUriString();
            URI location = new URI(baseUrl + "/api/places/" + id + "/image");
            Resource file = new InputStreamResource(imageFile.getInputStream());
            return ResponseEntity.ok().header(HttpHeaders.CONTENT_TYPE,  "image/" + imageFile.getOriginalFilename().split("\\.")[1], HttpHeaders.CONTENT_LOCATION, location.toString())
                    .contentLength(newPlace.getImageFile().length()).body(file);
        } else {
            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
        }
    }

}